refactor: select service with request
