+++ src/main/java/team/router/recycle/web/station/StationController.java
@@ -6,17 +6,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import team.router.recycle.domain.station.StationService;
+import team.router.recycle.domain.station.StationServiceFactory;
 
 @RestController
 @RequestMapping("/station")
 @RequiredArgsConstructor
 public class StationController {
 
-    private final StationService stationService;
+    private final StationServiceFactory stationServiceFactory;
 
     @PostMapping("/realtime")
     public ResponseEntity<?> getRealtimeStation(@RequestBody StationRealtimeRequest stationRealtimeRequest) {
-        return ResponseEntity.ok(stationService.getRealtimeStation(stationRealtimeRequest));
+        return ResponseEntity.ok(stationServiceFactory.getStationService(stationRealtimeRequest).getRealtimeStation(stationRealtimeRequest));
     }
 }
