+++ src/main/java/team/router/recycle/web/station/StationRealtimeRequest.java
@@ -4,4 +4,12 @@ public record StationRealtimeRequest(
         Double latitude,
         Double longitude
 ) {
+
+    public String getCity() {
+        if (latitude > 37) {
+            return "seoul";
+        } else {
+            return "daejeon";
+        }
+    }
 }
